+++ example/lib/api_keys.example.dart
@@ -1,5 +1,0 @@
-// TODO: Rename this file as api_keys.dart and insert your API Key & Secret.
-// You can obtain credentials from https://portal.iproov.com
-
-const apiKey = '< YOUR API KEY >';
-const secret = '< YOUR SECRET >';