+++ example/lib/credentials.example.dart
@@ -1,0 +1,9 @@
+// TODO: Rename this file as credentials.dart and update accordingly.
+// You can obtain credentials from https://portal.iproov.com
+//
+// IMPORTANT: This is provided for demo/example purposes only.
+//  You should *never* embed credentials in a production app!
+
+const baseUrl = 'https://eu.rp.secure.iproov.me/api/v2/';
+const apiKey = '< YOUR API KEY >';
+const secret = '< YOUR SECRET >';