+++ CHANGELOG.md
@@ -1,4 +1,25 @@
-# iProov Biometrics SDK
+# iProov Biometrics Flutter SDK
+
+## 1.0.0
+
+We're pleased to announce that the iProov Biometrics Flutter SDK is now production-ready!
+
+### Flutter
+
+* Added `floatingPromptEnabled` to `UiOptions`.
+* Renamed `footerTextColor` to `promptTextColor` in `UiOptions`.
+* Removed `font` and `fontResource` from `UiOptions`. Use `fontPath` instead, which is now cross-platform.
+
+### iOS
+
+* Upgraded SDK to [v9.3.0](https://github.com/iProov/ios/releases/tag/9.3.0).
+* Updated installation instructions for Cocoapods.
+* Added support for custom fonts.
+
+### Android
+
+* Upgraded SDK to [v7.2.0](https://github.com/iProov/android/releases/tag/v7.2.0).
+* Added support for custom fonts.
 
 ## 0.2.0
 
@@ -14,14 +35,14 @@
 
 ### iOS
 
-* Upgraded SDK to v9.2.0.
+* Upgraded SDK to [v9.2.0](https://github.com/iProov/ios/releases/tag/9.2.0).
 * Passing certificates directly as `List<int>` is now supported.
 * Error handling improvements.
 * `closeButtonImage` is now supported.
 
 ### Android
 
-* Upgraded SDK to v7.1.0.
+* Upgraded SDK to [v7.1.0](https://github.com/iProov/android/releases/tag/v7.1.0).
 * Passing certificates directly as `List<int>` is now supported.
 * Error handling improvements.
 
@@ -29,5 +50,6 @@
 
 Initial preview release
 
-* iOS SDK 9.0.1
-* Android SDK 7.0.3
+* iOS SDK [9.0.1](https://github.com/iProov/ios/releases/tag/9.0.1).
+* Android SDK [7.0.3](https://github.com/iProov/android/releases/tag/v7.0.3).
+*